Fixing for following bugs:
I get exceptions if I try to update a not available contactperson. This
happens if I enter two unknown data in the field and press update

If I pick a person, then press reset and afterwards click delete, the
app crashes

If I am correct, firstname and lastname of a contactperson, should not
be allowed to be null, yet the database accept whitespace as data derp.

I don't get any feedback when I add, delete or update a person, as in:
"Success, you have added person A"