ValidatieMessages
The validation messages get shown properly now if you start the application, and the odd behavior with input is gone. The cause was due to the message conflicting with the mahapps layer, they exist, but get overlapped. The AdornerDecorator acts as a sort z-index and enforces that they get shown.